+++ smooth-doc/gatsby-browser.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import { RootWrapper } from './src/components/RootWrapper'
 import { PageWrapper } from './src/components/PageWrapper'
+import Prism from 'prism-react-renderer/prism';
 
 export const wrapRootElement = ({ element }) => {
   return <RootWrapper>{element}</RootWrapper>
@@ -9,3 +10,6 @@ export const wrapRootElement = ({ element }) => {
 export const wrapPageElement = ({ element, props }) => {
   return <PageWrapper props={props}>{element}</PageWrapper>
 }
+
+(typeof global !== 'undefined' ? global : window).Prism = Prism;
+require('prismjs/components/prism-java');