+++ smooth-doc/gatsby-config.js
@@ -7,6 +7,19 @@ function getLogoPath() {
     : `${__dirname}/images/logo-manifest.png`
 }
 
+const visit = require('unist-util-visit')
+
+/** @type {import('unified').Plugin<Array<void>, import('hast').Root>} */
+function rehypeMetaAsAttributes() {
+  return (tree) => {
+    visit(tree, 'element', (node) => {
+      if (node.tagName === 'code' && node.data && node.data.meta) {
+        node.properties.meta = node.data.meta
+      }
+    })
+  }
+}
+
 /**
  * Theme configuration.
  * @param {object} options
@@ -15,7 +28,7 @@ function getLogoPath() {
  * @param {string} [options.siteUrl]
  * @param {string} [options.shortName]
  * @param {string[]} [options.sections]
- * @param {{ title: string, url: string }[]} [options.nav]
+ * @param {{ title: string, url: string }[]} [options.navItems]
  * @param {string} [options.baseDirectory]
  * @param {string} [options.twitterAccount]
  * @param {string} [options.githubRepositoryURL]
@@ -23,10 +36,10 @@ function getLogoPath() {
  * @param {string} [options.githubDefaultBranch]
  * @param {string} [options.author]
  * @param {string} [options.carbonAdsURL]
- * @param {{ apiKey: string, indexName: string }} [options.docSearch]
+ * @param {{ apiKey: string, indexName: string, appId: string }} [options.docSearch]
  * @param {object} [options.sitemap]
  */
-module.exports = (options) => {
+module.exports = function config(options) {
   const siteUrl = getSiteUrl(options)
   const logoPath = getLogoPath()
 
@@ -51,17 +64,28 @@ module.exports = (options) => {
           modules: ['smooth-doc'],
         },
       },
-      'gatsby-plugin-styled-components',
+      {
+        resolve: 'gatsby-plugin-styled-components',
+        options: {
+          topLevelImportPaths: ['@xstyled/styled-components'],
+        },
+      },
       'gatsby-remark-images',
       'gatsby-plugin-catch-links',
       {
         resolve: 'gatsby-plugin-mdx',
         options: {
+          extensions: [`.mdx`, `.md`],
+          mdxOptions: {
+            remarkPlugins: [
+              // Add GitHub Flavored Markdown (GFM) support
+              require(`remark-gfm`),
+            ],
+            rehypePlugins: [rehypeMetaAsAttributes],
+          },
           gatsbyRemarkPlugins: [
             {
-              resolve: require.resolve(
-                './src/plugins/gatsby-remark-autolink-headers',
-              ),
+              resolve: 'gatsby-remark-mermaid'
             },
             {
               resolve: `gatsby-remark-images`,
@@ -69,18 +93,13 @@ module.exports = (options) => {
                 maxWidth: 1200,
               },
             },
+            { resolve: 'gatsby-remark-autolink-headers' },
           ],
         },
       },
       'gatsby-transformer-sharp',
       'gatsby-plugin-sharp',
       'gatsby-plugin-react-helmet',
-      {
-        resolve: require.resolve(
-          './src/plugins/gatsby-remark-autolink-headers',
-        ),
-      },
-
       // Source
       {
         resolve: 'gatsby-source-filesystem',
@@ -99,7 +118,15 @@ module.exports = (options) => {
       {
         resolve: 'gatsby-source-filesystem',
         options: {
+          path: `./pages/docs`,
+          name: 'doc',
+        },
+      },
+      {
+        resolve: 'gatsby-source-filesystem',
+        options: {
           path: `./pages`,
+          ignore: ['**/docs/**'],
           name: 'page',
         },
       },
@@ -129,7 +156,7 @@ module.exports = (options) => {
       {
         resolve: 'gatsby-plugin-robots-txt',
         options: {
-          resolveEnv: () => process.env.NETLIFY_ENV || process.env.NODE_ENV,
+          resolveEnv: () => process.env.CONTEXT || process.env.NODE_ENV,
           env: {
             production: {
               policy: [{ userAgent: '*' }],