+++ smooth-doc/gatsby-node.js
@@ -14,6 +14,7 @@ function createSchemaCustomization({ actions }) {
     type AlgoliaDocSearchMetadata {
       apiKey: String!
       indexName: String!
+      appId: String
     }
 
     type SiteSiteMetadata {
@@ -31,6 +32,7 @@ function createSchemaCustomization({ actions }) {
 
     type MdxFrontmatter {
       title: String!
+      description: String
       slug: String
       section: String
       order: Int
@@ -57,7 +59,7 @@ async function onPreBootstrap(options) {
 function onCreateMdxNode({ node, getNode, actions }, options) {
   const { createNodeField } = actions
   const slug = node.frontmatter.slug || createFilePath({ node, getNode })
-  const pageType = /\/pages\/docs\//.test(node.fileAbsolutePath)
+  const pageType = /\/pages\/docs\//.test(node.internal.contentFilePath)
     ? 'doc'
     : 'page'
 
@@ -130,12 +132,15 @@ function onCreateMdxNode({ node, getNode, actions }, options) {
       baseDirectory,
       githubDocRepositoryURL,
       githubRepositoryURL,
-      githubDefaultBranch = 'master',
+      githubDefaultBranch = 'main',
     } = options
     const repositoryURL = githubDocRepositoryURL || githubRepositoryURL
     if (!baseDirectory || !repositoryURL) return ''
-    const relativePath = node.fileAbsolutePath.replace(baseDirectory, '')
-    return `${repositoryURL}/edit/${githubDefaultBranch}${relativePath}`
+    const relativePath = node.internal.contentFilePath.replace(
+      baseDirectory,
+      '',
+    )
+    return `${repositoryURL}edit/${githubDefaultBranch}${relativePath}`
   }
 
   createNodeField({
@@ -146,7 +151,7 @@ function onCreateMdxNode({ node, getNode, actions }, options) {
 }
 
 function onCreateNode(...args) {
-  if (args[0].node.internal.type === `Mdx`) {
+  if (args[0].node.internal.type === 'Mdx') {
     onCreateMdxNode(...args)
   }
 }
@@ -165,6 +170,14 @@ async function createPages({ graphql, actions, reporter }) {
               pageType
               redirect
             }
+            parent {
+              ... on File {
+                sourceInstanceName
+              }
+            }
+            internal {
+              contentFilePath
+            }
           }
         }
       }
@@ -173,30 +186,87 @@ async function createPages({ graphql, actions, reporter }) {
 
   if (errors) {
     reporter.panicOnBuild(`Error while running GraphQL query.`)
+    console.log(errors)
     return
   }
 
+  const filteredEdges = data.allMdx.edges.filter((edge) => {
+    if (edge.node.parent.sourceInstanceName === 'default-page') {
+      const { slug } = edge.node.fields
+      const hasCustom404 = data.allMdx.edges.find(
+        (_edge) => edge !== _edge && _edge.node.fields.slug === slug,
+      )
+      return !hasCustom404
+    }
+    return true
+  })
+
   // Create pages
-  const { edges } = data.allMdx
-  edges.forEach(({ node }) => {
+  filteredEdges.forEach(({ node }) => {
     if (node.fields.redirect) {
       createRedirect({
         fromPath: node.fields.slug,
         toPath: node.fields.redirect,
         redirectInBrowser: true,
       })
+      return
     }
 
     createPage({
       path: node.fields.slug,
       component: path.resolve(
         __dirname,
-        `./src/templates/${node.fields.pageType}.js`,
+        `./src/templates/${node.fields.pageType}.js?__contentFilePath=${node.internal.contentFilePath}`,
       ),
       context: {
         id: node.id,
+        frontmatter: node.frontmatter,
+        contentFilePath: node.internal.contentFilePath,
       },
     })
+  })
+}
+
+const pluginOptionsSchema = (/** @type {{ Joi: import('joi') }} */ { Joi }) => {
+  return Joi.object({
+    // Validate that the anonymize option is defined by the user and is a boolean
+    name: Joi.string().required(),
+    description: Joi.string().required(),
+    siteUrl: Joi.string(),
+    shortName: Joi.string(),
+    sections: Joi.array().items(Joi.string()),
+    navItems: Joi.array().items(
+      Joi.object({
+        title: Joi.string().required(),
+        url: Joi.string().required(),
+      }),
+    ),
+    baseDirectory: Joi.string(),
+    twitterAccount: Joi.string(),
+    githubRepositoryURL: Joi.string(),
+    githubDocRepositoryURL: Joi.string(),
+    githubDefaultBranch: Joi.string(),
+    author: Joi.string(),
+    carbonAdsURL: Joi.string(),
+    docSearch: Joi.object({
+      apiKey: Joi.string().required(),
+      indexName: Joi.string().required(),
+      appId: Joi.string(),
+    }),
+    sitemap: Joi.object(),
+  })
+}
+
+const onCreateWebpackConfig = ({ stage, rules, loaders, plugins, actions }) => {
+  actions.setWebpackConfig({
+    module: {
+      rules: [
+        {
+          test: /\.mdx$/,
+          use: require.resolve('@mdx-js/loader'),
+        },
+      ],
+    },
   })
 }
 
@@ -205,4 +275,6 @@ module.exports = {
   onPreBootstrap,
   onCreateNode,
   createPages,
+  pluginOptionsSchema,
+  onCreateWebpackConfig,
 }