+++ smooth-doc/src/components/AppNav.js
@@ -35,10 +35,11 @@ export function AppNav() {
         {data.site.siteMetadata.githubRepositoryURL ? (
           <NavListItem>
             <NavLink
-              forwardedAs="a"
+              as="a"
               href={data.site.siteMetadata.githubRepositoryURL}
               target="_blank"
               rel="noopener noreferrer"
+              aria-label="Github"
             >
               <RiGithubFill style={{ width: 24, height: 24 }} />
             </NavLink>
@@ -47,8 +48,9 @@ export function AppNav() {
         {data.site.siteMetadata.twitterAccount ? (
           <NavListItem>
             <NavLink
-              forwardedAs="a"
+              as="a"
               href={`https://twitter.com/${data.site.siteMetadata.twitterAccount}`}
+              aria-label="Twitter"
               target="_blank"
               rel="noopener noreferrer"
             >
@@ -57,7 +59,7 @@ export function AppNav() {
           </NavListItem>
         ) : null}
         <NavListItem>
-          <NavLink forwardedAs={ColorModeSwitcher} />
+          <NavLink aria-label="Switch Theme" as={ColorModeSwitcher} />
         </NavListItem>
       </NavList>
     </Nav>