+++ smooth-doc/src/components/DocSearch.js
@@ -1,82 +1,121 @@
-import React from 'react'
-import { Box, createGlobalStyle, th } from '@xstyled/styled-components'
-import { RiSearchEyeLine } from 'react-icons/ri'
+import * as React from 'react'
+import { createPortal } from 'react-dom'
+import styled, { x, createGlobalStyle } from '@xstyled/styled-components'
+import { DocSearchModal, useDocSearchKeyboardEvents } from '@docsearch/react'
+import { RiSearchLine } from 'react-icons/ri'
 import { Input, InputGroup, InputGroupIcon } from './Input'
 
-const DocSearchStyle = createGlobalStyle`
-  .algolia-autocomplete .algolia-docsearch-suggestion--highlight {
-    color: doc-search-suggestion-highlight-on-background;
-    background-color: doc-search-suggestion-highlight-background;
-  }
+require('@docsearch/css')
 
-  .algolia-autocomplete .algolia-docsearch-suggestion--text .algolia-docsearch-suggestion--highlight {
-    box-shadow: inset 0 -2px 0 0 ${th.color(
-      'doc-search-suggestion-content-underline',
-    )};
-  }
-
-  .algolia-autocomplete .ds-cursor .algolia-docsearch-suggestion--content {
-    background-color: doc-search-suggestion-content-background !important;
-  }
-
-  .algolia-autocomplete .ds-dropdown-menu [class^=ds-dataset-] {
-    background-color: background;
-    border-color: layout-border;
-  }
-
-  .algolia-autocomplete .ds-dropdown-menu:before {
-    background-color: background;
-    border-color: layout-border;
-  }
-
-  .algolia-autocomplete .algolia-docsearch-suggestion {
-    background-color: background;
-    color: on-background;
-  }
-
-  .algolia-autocomplete .algolia-docsearch-suggestion--category-header {
-    border-color: layout-border;
-    color: on-background-light;
-  }
-
-  .algolia-autocomplete .algolia-docsearch-suggestion--title {
-    color: on-background;
-  }
-
-  .algolia-autocomplete .algolia-docsearch-suggestion--subcategory-column {
-    color: on-background-light;
-  }
-
-  .algolia-autocomplete .algolia-docsearch-suggestion--text {
-    color: on-background-light;
+const GlobalStyle = createGlobalStyle`
+  /* Darkmode */
+  body.xstyled-color-mode-dark {
+    --docsearch-text-color: rgb(245, 246, 247);
+    --docsearch-container-background: rgba(9, 10, 17, 0.8);
+    --docsearch-modal-background: rgb(21, 23, 42);
+    --docsearch-modal-shadow: inset 1px 1px 0 0 rgb(44, 46, 64),
+      0 3px 8px 0 rgb(0, 3, 9);
+    --docsearch-searchbox-background: rgb(9, 10, 17);
+    --docsearch-searchbox-focus-background: #000;
+    --docsearch-hit-color: rgb(190, 195, 201);
+    --docsearch-hit-shadow: none;
+    --docsearch-hit-background: rgb(9, 10, 17);
+    --docsearch-key-gradient: linear-gradient(
+      -26.5deg,
+      rgb(86, 88, 114) 0%,
+      rgb(49, 53, 91) 100%
+    );
+    --docsearch-key-shadow: inset 0 -2px 0 0 rgb(40, 45, 85),
+      inset 0 0 1px 1px rgb(81, 87, 125), 0 2px 2px 0 rgba(3, 4, 9, 0.3);
+    --docsearch-footer-background: rgb(30, 33, 54);
+    --docsearch-footer-shadow: inset 0 1px 0 0 rgba(73, 76, 106, 0.5),
+      0 -4px 8px 0 rgba(0, 0, 0, 0.2);
+    --docsearch-logo-color: rgb(255, 255, 255);
+    --docsearch-muted-color: rgb(127, 132, 151);
   }
 `
 
-export function DocSearch({ apiKey, indexName }) {
-  React.useEffect(() => {
-    window.docsearch({
-      apiKey,
-      indexName,
-      inputSelector: '#algolia-docsearch-input',
-      debug: true,
-    })
-  }, [apiKey, indexName])
+const Kbd = styled.kbd`
+  border: 1;
+  border-color: control-border;
+  margin-right: 1;
+  background-color: control-background;
+  text-align: center;
+  padding: 0;
+  display: inline-flex;
+  justify-content: center;
+  font-size: 0.8em;
+  line-height: 1.2;
+  font-family: sans-serif;
+  border-radius: base;
+  min-width: 1.5em;
+`
+
+export function DocSearch({ apiKey, indexName, appId }) {
+  const searchButtonRef = React.useRef(null)
+  const [isOpen, setIsOpen] = React.useState(false)
+  const [initialQuery, setInitialQuery] = React.useState(null)
+
+  const onOpen = React.useCallback(() => {
+    setIsOpen(true)
+  }, [setIsOpen])
+
+  const onClose = React.useCallback(() => {
+    setIsOpen(false)
+  }, [setIsOpen])
+
+  const onInput = React.useCallback(
+    (event) => {
+      setIsOpen(true)
+      setInitialQuery(event.key)
+    },
+    [setIsOpen, setInitialQuery],
+  )
+
+  useDocSearchKeyboardEvents({
+    isOpen,
+    onOpen,
+    onClose,
+    onInput,
+    searchButtonRef,
+  })
 
   return (
     <>
-      <DocSearchStyle />
-      <Box>
-        <InputGroup>
+      <GlobalStyle />
+      <div>
+        <InputGroup ref={searchButtonRef} as="button" onClick={() => onOpen()}>
           <InputGroupIcon>
-            <RiSearchEyeLine />
+            <RiSearchLine />
           </InputGroupIcon>
-          <Input
-            id="algolia-docsearch-input"
-            type="search"
-            placeholder="Search..."
-          />
+          <Input disabled type="search" placeholder="Search..." />
+          <x.div
+            position="absolute"
+            top="50%"
+            right={0}
+            transform="translateY(-50%)"
+            display="inline-flex"
+            pointerEvents="none"
+            userSelect="none"
+          >
+            <Kbd>⌘</Kbd>
+            <Kbd>K</Kbd>
+          </x.div>
         </InputGroup>
-      </Box>
+      </div>
+
+      {isOpen &&
+        createPortal(
+          <DocSearchModal
+            apiKey={apiKey}
+            indexName={indexName}
+            appId={appId}
+            onClose={onClose}
+            initialScrollY={window.scrollY}
+            initialQuery={initialQuery}
+          />,
+          document.body,
+        )}
     </>
   )
 }