+++ smooth-doc/src/components/MDX.js
@@ -5,19 +5,19 @@ import { CarbonAd } from './CarbonAd'
 import { Table, TableContainer } from './Table'
 
 function transformCode({ children, className, ...props }) {
-  const lang = className && className.split('-')[1]
-  return (
-    <Code lang={lang} {...props}>
-      {children}
-    </Code>
-  )
+    const lang = className && className.split('-')[1]
+    return (
+      <Code lang={lang} {...props}>
+        {children}
+      </Code>
+    )
 }
 
 function getCodeChild(children) {
   const childrenArray = React.Children.toArray(children)
   if (childrenArray.length !== 1) return null
   const [firstChild] = childrenArray
-  if (firstChild.props.mdxType !== 'code') return null
+  if (firstChild.type !== 'code') return null
   return firstChild
 }