+++ smooth-doc/src/components/TableOfContents.js
@@ -1,4 +1,5 @@
 import React from 'react'
+// eslint-disable-next-line import/no-unresolved
 import { useLocation } from '@reach/router'
 import styled, { th } from '@xstyled/styled-components'
 
@@ -160,7 +161,7 @@ export function TableOfContents() {
   if (!headings.length) return null
   return (
     <TocContainer ref={ref}>
-      <h4>このページの概要</h4>
+      <h4>On this page</h4>
       <ul>
         {headings.map((heading, i) =>
           heading.url ? (